+++ modules/apps/commerce/commerce-inventory-service/src/main/java/com/liferay/commerce/inventory/internal/type/MoveQuantityCommerceInventoryAuditTypeImpl.java
@@ -23,6 +23,8 @@ import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.service.UserLocalService;
 
+import java.math.BigDecimal;
+
 import java.util.Locale;
 import java.util.Map;
 
@@ -57,7 +59,7 @@ public class MoveQuantityCommerceInventoryAuditTypeImpl
 	}
 
 	@Override
-	public String formatQuantity(int quantity, Locale locale) {
+	public String formatQuantity(BigDecimal quantity, Locale locale) {
 		return String.valueOf(quantity);
 	}
