+++ modules/apps/commerce/commerce-inventory-web/src/main/java/com/liferay/commerce/inventory/web/internal/model/Replenishment.java
@@ -14,6 +14,8 @@
 
 package com.liferay.commerce.inventory.web.internal.model;
 
+import java.math.BigDecimal;
+
 /**
  * @author Luca Pellizzon
  * @author Alessio Antonio Rendina
@@ -22,7 +24,7 @@ public class Replenishment {
 
 	public Replenishment(
 		long commerceInventoryReplenishmentItemId, String warehouse,
-		String date, int quantity) {
+		String date, BigDecimal quantity) {
 
 		_commerceInventoryReplenishmentItemId =
 			commerceInventoryReplenishmentItemId;
@@ -39,7 +41,7 @@ public class Replenishment {
 		return _date;
 	}
 
-	public int getQuantity() {
+	public BigDecimal getQuantity() {
 		return _quantity;
 	}
 
@@ -49,7 +51,7 @@ public class Replenishment {
 
 	private final long _commerceInventoryReplenishmentItemId;
 	private final String _date;
-	private final int _quantity;
+	private final BigDecimal _quantity;
 	private final String _warehouse;
 
 }
