+++ modules/apps/commerce/commerce-service/src/main/java/com/liferay/commerce/internal/inventory/CPDefinitionOptionValueRelCommerceInventoryCheckerImpl.java
@@ -19,6 +19,8 @@ import com.liferay.commerce.product.model.CPDefinitionOptionValueRel;
 import com.liferay.portal.kernel.model.BaseModel;
 import com.liferay.portal.kernel.util.Validator;
 
+import java.math.BigDecimal;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -64,7 +66,7 @@ public class CPDefinitionOptionValueRelCommerceInventoryCheckerImpl
 
 		return isAvailable(
 			cpDefinitionOptionValueRel.fetchCPInstance(),
-			cpDefinitionOptionValueRel.getQuantity());
+			BigDecimal.valueOf(cpDefinitionOptionValueRel.getQuantity()));
 	}
 
 }
